requirements.txt
tensorflow
numpy
pandas
matplotlib
scikit-learn
seaborn
streamlit
pillow
